SLOPE: Sorted L1 Penalized Estimation

Efficient implementations for Sorted L-One Penalized Estimation (SLOPE): generalized linear models regularized with the sorted L1-norm (Bogdan et al. 2015). Supported models include ordinary least-squares regression, binomial regression, multinomial regression, and Poisson regression. Both dense and sparse predictor matrices are supported. In addition, the package features predictor screening rules that enable fast and efficient solutions to high-dimensional problems.

See also

Author

Maintainer: Johan Larsson johanlarsson@outlook.com (ORCID)

Authors:

Jonas Wallin jonas.wallin@stat.lu.se (ORCID)

Malgorzata Bogdan (ORCID)

Ewout van den Berg

Chiara Sabatti

Emmanuel Candes

Evan Patterson

Weijie Su

Jakub Kała

Krystyna Grzesiak

Mathurin Massias

Quentin Klopfenstein

Michal Burdukiewicz (ORCID)

Other contributors:

Jerome Friedman (code adapted from 'glmnet') [contributor]

Trevor Hastie (code adapted from 'glmnet') [contributor]

Rob Tibshirani (code adapted from 'glmnet') [contributor]

Balasubramanian Narasimhan (code adapted from 'glmnet') [contributor]

Noah Simon (code adapted from 'glmnet') [contributor]

Junyang Qian (code adapted from 'glmnet') [contributor]