Abalone

This data set contains observations of abalones, the common name for any of a group of sea snails. The goal is to predict the age of an individual abalone given physical measurements such as sex, weight, and height.

Usage

abalone

Format

A list with two items representing 211 observations from 9 variables

sex

sex of abalone, 1 for female

infant

indicates that the person is an infant

length

longest shell measurement in mm

diameter

perpendicular to length in mm

height

height in mm including meat in shell

weight_whole

weight of entire abalone

weight_shucked

weight of meat

weight_viscera

weight of viscera

weight_shell

weight of shell

rings

rings. +1.5 gives the age in years

Source

Pace, R. Kelley and Ronald Barry, Sparse Spatial Autoregressions, Statistics and Probability Letters, 33 (1997) 291-297.

Details

Only a stratified sample of 211 rows of the original data set are used here.

See also