Bodyfat

The response (y) corresponds to estimates of percentage of body fat from application of Siri's 1956 equation to measurements of underwater weighing, as well as age, weight, height, and a variety of body circumference measurements.

Usage

bodyfat

Format

A list with two items representing 252 observations from 14 variables

age

age (years)

weight

weight (lbs)

height

height (inches)

neck

neck circumference (cm)

chest

chest circumference (cm)

abdomen

abdomen circumference (cm)

hip

hip circumference (cm)

thigh

thigh circumference (cm)

knee

knee circumference (cm)

ankle

ankle circumference (cm)

biceps

biceps circumference (cm)

forearm

forearm circumference (cm)

wrist

wrist circumference (cm)

Source

http://lib.stat.cmu.edu/datasets/bodyfat

https://www.csie.ntu.edu.tw/~cjlin/libsvmtools/datasets/regression.html

See also