Model deviance

Model deviance

# S3 method for SLOPE
deviance(object, ...)

Arguments

object
an object of class 'SLOPE'.

...
ignored

Value

For Gaussian models this is twice the residual sums of squares. For all other models, two times the negative loglikelihood is returned.

Examples

fit <- SLOPE(heart$x, heart$y, family = "binomial")
deviance(fit)

#>  [1] 374.2995 374.2995 335.5373 320.6967 307.3978 295.8861 284.7133 274.4445
#>  [9] 265.4250 257.5443 250.5870 244.1222 238.4380 232.3275 226.2262 220.7925
#> [17] 215.8417 211.5264 207.4409 203.8190 200.6200 197.7982 195.3319 193.2309
#> [25] 191.3185 189.6012 188.0917 186.7159 185.5211 184.4934 183.5830 182.7930
#> [33] 182.1122 181.5189 181.0046 180.5680 179.8512 178.8020 177.5680 176.5844
#> [41] 175.8327 175.0133 174.4435 173.7988 173.1948 172.8435 172.3317 171.8640
#> [49] 171.4949 171.2087 170.9059 170.6504 170.4349 170.2533 170.0991 169.9663
#> [57] 169.8328 169.6833 169.5587 169.4520 169.3632 169.2861 169.2233 169.1706
#> [65] 169.1253 169.0879 169.0560 169.0294 169.0071 168.9886 168.9731 168.9611
#> [73] 168.9501 168.9410 168.9337 168.9274 168.9222 168.9178 168.9141 168.9111
#> [81] 168.9086 168.9065 168.9048 168.9033