Model deviance

Model deviance

# S3 method for SLOPE
deviance(object, ...)

Arguments

object
an object of class 'SLOPE'.

...
ignored

Value

For Gaussian models this is twice the residual sums of squares. For all other models, two times the negative loglikelihood is returned.

Examples

fit <- SLOPE(abalone$x, abalone$y, family = "poisson")
deviance(fit)

#>  [1] 6822.7872 1026.6163 1018.0849 1010.0639 1003.8413  998.6156  993.9010
#>  [8]  989.1671  985.8107  982.9873  980.2945  977.8246  975.9313  974.3124
#> [15]  972.9612  971.8876  970.9016  970.0228  969.2916  968.6316  968.0704
#> [22]  966.8109  963.6936  961.0821  958.9515  957.2989  955.2906  953.6264
#> [29]  952.3298  951.1846  950.1994  949.3810  948.7004  948.1341  947.6637
#> [36]  947.2721  946.8823  946.5156  946.2148  945.7857  945.3086  944.7726
#> [43]  944.3939  943.8796  943.5149  943.1791  942.8186  942.5244  942.3303
#> [50]  942.1444  941.9637  941.8125  941.6787  941.5672  941.4741  941.3968
#> [57]  941.3324  941.2788  941.2341  941.1974  941.1671  941.1413  941.1198
#> [64]  941.1019  941.0871  941.0747  941.0645  941.0571