Model deviance

Model deviance

Usage

# S3 method for class 'SLOPE'
deviance(object, ...)

Arguments

object

an object of class 'SLOPE'.

...

ignored

Value

For Gaussian models this is twice the residual sums of squares. For all other models, two times the negative loglikelihood is returned.

See also

Examples

fit <- SLOPE(heart$x, heart$y, family = "binomial")
deviance(fit)
#>  [1] 1.3739232 1.2937213 1.2251813 1.1670235 1.1161943 1.0717583 1.0275800
#>  [8] 0.9889878 0.9549637 0.9250257 0.8985209 0.8753882 0.8552442 0.8353384
#> [15] 0.8138016 0.7943370 0.7767139 0.7611075 0.7459662 0.7323982 0.7197063
#> [22] 0.7083896 0.6986941 0.6897190 0.6822010 0.6752986 0.6691959 0.6637329
#> [29] 0.6587841 0.6545033 0.6506909 0.6473870 0.6445433 0.6420809 0.6400310
#> [36] 0.6381617 0.6365579 0.6350098 0.6336420 0.6324507 0.6314430 0.6305686
#> [43] 0.6298191 0.6291761 0.6286302 0.6281726 0.6277802 0.6274314 0.6271386
#> [50] 0.6268907 0.6266718 0.6264904 0.6263364 0.6262055 0.6260964 0.6260050
#> [57] 0.6259285 0.6258651 0.6258111 0.6257663 0.6257289 0.6256976 0.6256715
#> [64] 0.6256498 0.6256316 0.6256165 0.6256039 0.6255934 0.6255846 0.6255774
#> [71] 0.6255713