Ridges and Tops

Detects different types of Ridges and Tops, based on Silveira and Silveira (2020). The primary difference is the use of a rectangular local neighborhood shape instead of a circular one, in calculus of White Top Hat (WTH) (Rodriguez et al. 2002). For more satisfactory results, it is recommended to define the moving window based on the minimum mappable area.

Usage

ridges_tops(x, sp_range, type)

Arguments

x

Input DEM raster file.

sp_range

Numeric. Number of neighbor cells for multiscalar analysis.

type

Numeric. Each number represents a specific landform, for example: (1) Convex Hilltops and Interfluves, (2) Sharp Crests, and (3) Ridges

Examples

library(terra)
library(slope)
dem <- terra:rast(system.file("ex/elev.tif", package = "terra"))
#> Error: object 'terra' not found
r <- slope::ridges_tops(dem, 7, 1)
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'x' in selecting a method for function 'rast': error in evaluating the argument 'x' in selecting a method for function 'sources': object 'dem' not found
plot(r)
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'x' in selecting a method for function 'plot': object 'r' not found