Compute one of several loss metrics on a new data set

This function is a unified interface to return various types of loss for a model fit with SLOPE().

score(object, x, y, measure)

# S3 method for GaussianSLOPE
score(object, x, y, measure = c("mse", "mae"))

# S3 method for BinomialSLOPE
score(object, x, y, measure = c("mse", "mae", "deviance", "misclass", "auc"))

# S3 method for MultinomialSLOPE
score(object, x, y, measure = c("mse", "mae", "deviance", "misclass"))

# S3 method for PoissonSLOPE
score(object, x, y, measure = c("mse", "mae"))

Arguments

object
an object of class "SLOPE"

x
feature matrix

y
response

measure
type of target measure. "mse" returns mean squared error. "mae" returns mean absolute error, "misclass" returns misclassification rate, and "auc" returns area under the ROC curve.

Value

The measure along the regularization path depending on the value in measure.

Examples

x <- subset(infert, select = c("induced", "age", "pooled.stratum"))
y <- infert$case

fit <- SLOPE(x, y, family = "binomial")
score(fit, x, y, measure = "auc")

#>        p1        p2        p3        p4        p5        p6 
#> 0.5150055 0.4921504 0.4952903 0.5025192 0.5090909 0.5095290